moving to vite and tailwind
