+++ client/package.json
@@ -3,30 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^1.7.9",
+    "@tailwindcss/cli": "^4.1.3",
+    "@tailwindcss/vite": "^4.1.3",
     "bootstrap": "^5.3.3",
     "client": "file:",
     "cra-template": "1.2.0",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.9",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^7.1.5",
-    "react-scripts": "^5.0.1",
     "socket.io-client": "^4.8.1",
+    "tailwindcss": "^4.1.3",
     "web-vitals": "^4.2.4",
     "zustand": "^5.0.3"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "browserslist": {
     "production": [
@@ -39,5 +32,12 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@vitejs/plugin-react": "^4.3.4",
+    "autoprefixer": "^10.4.21",
+    "postcss": "^8.5.3",
+    "postcss-cli": "^11.0.1",
+    "vite": "^6.2.5"
   }
 }
