+++ client/src/components/WhisperTextModal.jsx
